add math.rs
