+++ src/lambda/mod.rs
@@ -3,6 +3,7 @@ mod evaluation_context;
 mod expression;
 mod graphql_operation_context;
 mod lambda;
+mod math;
 mod resolver_context_like;
 
 pub use eval::*;
@@ -10,4 +11,5 @@ pub use evaluation_context::EvaluationContext;
 pub(crate) use expression::*;
 pub use graphql_operation_context::GraphQLOperationContext;
 pub use lambda::Lambda;
+pub use math::*;
 pub use resolver_context_like::{EmptyResolverContext, ResolverContextLike};
